requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic-settings
python-dotenv
psycopg2-binary
google-genai
sentence-transformers